@@ -5,6 +5,7 @@
   "author": "",
   "private": true,
   "scripts": {
+    "dev": "vite",
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint"
@@ -12,39 +13,37 @@
   "dependencies": {
     "axios": "^0.21.1",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
+    "vue": "^3.2.25",
+    "vue-router": "^4.0.0",
+    "vuex": "^4.0.0"
   },
   "devDependencies": {
+    "@vitejs/plugin-vue": "^5.2.0",
+    "@vitejs/plugin-vue-jsx": "^4.1.0",
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
+    "vite": "^5.4.11",
+    "vite-plugin-compression": "^0.5.1",
     "vue-template-compiler": "^2.6.11"
   },
-
   "eslintConfig": {
-
     "root": true,
     "env": {
       "node": true
     },
-
     "extends": [
       "plugin:vue/essential",
       "eslint:recommended"
     ],
-
     "parserOptions": {
       "parser": "babel-eslint"
     },
-
     "rules": {}
   },
-
   "browserslist": [
     "> 1%",
     "last 2 versions",
